More Setup
Front-end setup using react
